MBR: Bootloader merging
Added support for merging the bootloader (commit) with an existing
partition table, DOS(MBR) format, and writing the changes back to disk.

Added integrity checking to warn if the merge causes potential
corruption of the partition table data. The user is free to disregard
this warning and commit the changes if they wish.

The integrity check ignore the 'Boot Signature' portion, since the disk
may or may not have the '55 AA' pattern set, while the bootloader most
likely does.